backend/requirements.txt
requests
fastapi[standard]
pydantic
dotenv-python